moved softwareproject out of the components folder
